update component demo
